// src/models/categoriesAndTags.d.ts
export interface Datum {
    id:         number;
    attributes: PurpleAttributes;
}

export interface PurpleAttributes {
    Title:       string;
    createdAt:   Date;
    updatedAt:   Date;
    publishedAt: Date;
    slug:        string;
    posts:       Posts;
}

export interface Posts {
    data: PostsDatum[];
}